+++ CHANGELOG.md
@@ -6,6 +6,11 @@ For more information about changelogs, check
 [Keep a Changelog](http://keepachangelog.com) and
 [Vandamme](http://tech-angels.github.io/vandamme).
 
+## 0.32.1 - 2017-08-14
+
+* [FEATURE] Add `Yt::ContentOwner#bulk_report_jobs`
+* [FEATURE] Add `Yt::BulkReportJob#bulk_reports`
+
 ## 0.32.0 - 2017-07-05
 
 **How to upgrade**